change to xsearch behavior when single result is found.
Previously, when a search returned a single result the xsearch UI
would return immediately.  Now it actually displays the result and
waits for user input.  This allows checking the actual context before
accepting.

The exception is that if the xsearch is initiated with a search term
we *do* wish to return immediately.  This allows the same interface to
be used for fetching a single known password entry.

Move test for single result out of _search(), and check only when
initiating.