xsearch: center window in some cases
There is something strange going on here.  The window centering seems
to work fine when invoked for null search results, but seems to cause
some sort of weird delay/breakage when called after filling the select
list.  I have no idea what the source of the problem is.  I'll just
add it uncommented after the null result situation and hope we later
figure out what the issue is.