use ENCRYPT_NO_COMPRESS if available
The CRIME attacks on TLS showed that repeated compression of a message
with a variant component can leak some information about the plaintext
if an attacker knows something about the variant.

While it seems unlikely that such an attack applies directly to
assword, it's possible that some use of assword in an automated
situation could fit this attack profile.

Upcoming versions of gpgme should support a GPGME_ENCRYPT_NO_COMPRESS
option.  python-gpgme should add support for it once gpgme exposes the
flag.  The use of this flag should invalidate any CRIME-style attacks.

See the IETF OpenPGP WG list for discussion about how CRIME might or
might not apply to:

https://www.ietf.org/mail-archive/web/openpgp/current/msg07252.html

Note that this fix will only take effect once the dependent libraries
are themselves updated to document and support the ENCRYPT_NO_COMPRESS
flag during encrypt() and encrypt_sign() functions.