gui: static method to reset signal for object
gobject doesn't seem to enable good tracking of signal callbacks, so we have
to do it manually.

If this new method is used to set a signal callback it removes any callbacks
that had been registered with this method, and then sets the new one.  The
signal callback handler ID is recorded as an attribute of the object being
connected to.

This seems overly hacky to me, but it more or less does the trick.