gui: add ability to replace/reset password
If the context entry matches an entry in the db, the button will present the
option to "Replace" the password of the matched context.  The matched context
will be cached, and the text entry will then be re-purposed for entry of a
replacement password.  The button will be reassigned to offer "Reset",
which will reset the password to a new random string.