store version number inside dedicated version.py file
This change removes our dependency on pkg_resources, and makes it
possible to run "./assword --version" from the source directly (this
is used by help2man in debian packaging at least).

The change fixes a bug where the debian package would presumably FTBFS
on a system that doesn't already have assword installed, and another
bug on systems where assword is installed, but the version number
written to the manpage would reflect the system version, and not the
version in the build tree.

this is method 3 of
https://packaging.python.org/en/latest/single_source_version/